change id to long

+++ src/main/java/com/hims/entity/repository/MasHospitalRepository.java
@@ -3,5 +3,5 @@ package com.hims.entity.repository;
 import com.hims.entity.MasHospital;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MasHospitalRepository extends JpaRepository<MasHospital, Integer> {
+public interface MasHospitalRepository extends JpaRepository<MasHospital, Long> {
 }
