+++ src/main/java/com/hims/entity/repository/MasMaritalStatusRepository.java
@@ -3,5 +3,5 @@ package com.hims.entity.repository;
 import com.hims.entity.MasMaritalStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MasMaritalStatusRepository extends JpaRepository<MasMaritalStatus, Integer> {
+public interface MasMaritalStatusRepository extends JpaRepository<MasMaritalStatus, Long> {
 }
