+++ src/main/java/com/hims/entity/repository/MasReligionRepository.java
@@ -3,5 +3,5 @@ package com.hims.entity.repository;
 import com.hims.entity.MasReligion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MasReligionRepository extends JpaRepository<MasReligion, Integer> {
+public interface MasReligionRepository extends JpaRepository<MasReligion, Long> {
 }
